Django_Intershala/requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
blinker==1.4
cachetools==4.1.1
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.6.0
Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
django-credit-cards==0.4.1
django-filter==2.4.0
django-localflavor==3.0.1
django-money==1.3.1
django-phone-field==1.8.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-sendsms==0.4
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
freeze==3.0
google-auth==1.23.0
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
importlib-metadata==3.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
kombu==5.0.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
prompt-toolkit==3.0.8
protobuf==3.14.0
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-stdnum==1.16
python3-openid==3.2.0
pytz==2021.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
win-inet-pton==1.1.0
wincertstore==0.2
zipp==3.4.0
